Refactoring-and-Merging
During the merging process, many files have changed and the operation of many functions has been altered. Where possible the code written in the Python is used.

The functions are called on the various pages of the FLASK UI.

While still lots to address, we have a MVP. Some issues may still be present on some pages during this process.

Not all functionality is yet in the UI either.

At this stage
1. system configuration is working (basic)
2. status is working
3. Reports are working

API and various capabilities still to get to MVP goal.